// ProTime-R3-ui/src/app/api/model/user.ts
/**
 * ProTimeR3
 * Project Time Recording Tool by Dominik Mathmann
 *
 * The version of the OpenAPI document: 3.2
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    id?: number;
    token?: string;
    user?: string;
    password?: string;
}